Update
1. 인벤토리에 골드 표시
2. 아이템 정보에 되팔기 가격 표시
3. 경험치 및 골드에 콤마 표기법 적용
4. 스킬창 및 아이템창에 항상 고정된 Text는 별도로 분리.

+++ Assets/KWS/Scripts/Manager/RaidManager.cs
@@ -15,17 +15,14 @@ public class RaidManager : MonoBehaviour
         public Transform[] monster_spawn_position;
     }
 
-    [SerializeField]
-    private SpawnInfo[] spawn_infos;
+    [SerializeField] private SpawnInfo[] spawn_infos;
 
     private int spawn_info_idx;
     private int max_monster;
     private int now_monster;
 
-    [SerializeField]
-    private float compensation_exp;
-    [SerializeField]
-    private float compensation_gold;
+    [SerializeField] private int compensation_exp;
+    [SerializeField] private int compensation_gold;
 
     [SerializeField]
     private BattleResultUIManager battle_result_ui_manager;
